% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luc.R
\name{luc_boxplot}
\alias{luc_boxplot}
\title{Period Boxplot}
\usage{
luc_boxplot(
  df = luc_output,
  title = "Period (hours) distribution of genotypes",
  order = NA
)
}
\arguments{
\item{df}{Data frame containing circadian period and genotype data for
individual regions or entries, defaulting to \code{luc_output}. The function
assumes the existence of \code{Genotype} and \code{Period} columns within the data
frame.}

\item{title}{Title of the plot as a character string.}

\item{order}{Order of genotypes or families within the plot, from top to
bottom, as a vector of character strings. \code{luc_scatter()} uses this
argument to assign levels to the \code{Genotype} column using \code{factor()}.
Accordingly, this argument must contain all values found within the
\code{Genotype} column, matched for case, or else some values found in
\code{Genotype} will be erased. If left as \code{NA}, the default value,
\code{luc_scatter()} facets the scatterplot using the default order of values
found in the \code{Genotype} column.}
}
\value{
A plot containing multiple boxplots, printed to the console and
written to the working directory as a .png file.
}
\description{
Creates boxplots of circadian period for families or genotypes
using functions found in \code{ggplot2} and combines them into a single plot.
\code{luc_boxplot()} is designed to work with the data found in \code{luc_output},
the data frame produced by \code{luc_organize()}, although any dataset
containing circadian period and genotype data for individual regions or
entries may be used. The plot is printed to the console and written to the
working directory as a .png file. Data on no more than 20 families or
genotypes may be plotted using this function.
}
\examples{
\dontrun{
luc_boxplot(df = luc_output,
     title = "Dull 1010 Period Distribution",
     order = c("Col", "80.8", "90.9"))}

}
