% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luc.R
\name{luc_compare}
\alias{luc_compare}
\title{Compare Circadian Periods}
\usage{
luc_compare(df = luc_output, names)
}
\arguments{
\item{df}{Data frame containing circadian period and genotype data for
individual regions or entries, defaulting to \code{luc_output}. The function
assumes the existence of \code{Genotype} and \code{Period} columns within the data
frame. This data frame may contain data on more genotypes or families than
the two compared by this function.}

\item{names}{Names of two families or genotypes found within the \code{Genotype}
column, as a vector of exactly two character strings.}
}
\value{
A list containing the results of the t-test, written to the global
environment. A summary of results is also printed to the console.
}
\description{
Performs two-sample t-test for difference in mean circadian
period between two different families or genotypes using \code{t.test()}.
\code{luc_compare()} is designed to work with the data found in \code{luc_output},
the data frame produced by \code{luc_organize()}, although any dataset
containing circadian period and genotype data for individual regions or
entries may be used. Results of the t-test are written to the global
environment.
}
\examples{
\dontrun{
luc_compare(df = luc_output, names = c("Col", "80.1"))}

}
