% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luc.R
\name{luc_organize}
\alias{luc_organize}
\title{Organize and Combine imp_detrended_results Files}
\usage{
luc_organize(names, cnd_sheet, zt_lum_sheet)
}
\arguments{
\item{names}{Vector of character strings containing the names of
imp_detrended_results files (in the general format "myfile.xls") to be
loaded and organized by the function. The function assumes that these files
are Excel files.}

\item{cnd_sheet}{Name of the .cnd sheet within each imp_detrended_results
file as a character string. This sheet contains information on the period,
amplitude, phase, and RAE of regions for which period could be assigned. As
this sheet is automatically generated by MetaMorph, the name of this sheet
is typically the same in each file. If the names of these sheets are not
the same, rename the .cnd sheet in each file so that their names are
identical.}

\item{zt_lum_sheet}{Name of the sheet within each imp_detrended_results file
containing observations on the luminescence values at each ZT time for
individual regions, as a character string. If the names of these sheets are
not the same, rename this sheet in each file so that their names are
identical. This function assumes several characteristics of data within
these sheets, including the following: each row contains data from a
particular region; the first five columns include \code{File}, \verb{Plate No},
\code{Genotype}, \code{Well / Region}, and \verb{ZT Time}; the \code{Well / Region} column
contains observations of the general form "ROI '#' Ellipse (Integrated)";
and every column beyond and including the sixth column is a ZT time at
which regions were imaged and the observations within are specific
luminescence values for individual regions.}
}
\value{
A tibble written to the global environment and to the working
directory as a .csv file.
}
\description{
Loads, organizes, and combines imp_detrended_results files
produced by MetaMorph and BRASS per the workflow recommended by Cassandra
Baker for analyzing luciferase reporter assays on Arabidopsis thaliana.
Combined data is written to the global environment as a tibble and to the
working directory as a .csv file.
}
\examples{
\dontrun{
luc_organize(names = c("Dull503_1-100.xls", "Dull503_101-200.xls"),
     cnd_sheet = "Dull503_.cnd",
     zt_lum_sheet = "Sheet1")}

}
