% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luc.R
\name{luc_scatter}
\alias{luc_scatter}
\title{Period/RAE Scatterplot}
\usage{
luc_scatter(
  df = luc_output,
  title = "Period (hours) vs. RAE",
  order = NA,
  labels = T
)
}
\arguments{
\item{df}{Data frame containing circadian period, RAE, region, and genotype
data for individual regions or entries, defaulting to \code{luc_output}. The
function assumes the existence of \code{Genotype}, \code{RAE}, \code{Region}, and \code{Period}
columns within the data frame.}

\item{title}{Title of the plot as a character string.}

\item{order}{Order of genotypes or families within the faceted scatterplot as
a vector of character strings. \code{luc_scatter()} uses this argument to assign
levels to the \code{Genotype} column using \code{factor()}. Accordingly, this
argument must contain all values found within the \code{Genotype} column,
matched for case, or else some values found in \code{Genotype} will be erased.
If left as \code{NA}, the default value, \code{luc_scatter()} facets the scatterplot
using the default order of values found in the \code{Genotype} column.}

\item{labels}{Logical indicating whether to include or exclude region labels
from the plot.}
}
\value{
A faceted scatterplot, printed to the console and written to the
working directory as a .png file.
}
\description{
Plots period and RAE of individual regions or entries as a
scatterplot faceted by family or genotype, using functions found in
\code{ggplot2}.\code{luc_scatter()} is designed to work with the data found in
\code{luc_output}, the data frame produced by \code{luc_organize()}, although any
dataset containing circadian period, RAE, region, and genotype data for
individual regions or entries may be used. The plot is printed to the
console and written to the working directory as a .png file. No more than
20 facets may be plotted using this function.
}
\examples{
\dontrun{
luc_scatter(df = luc_output,
     title = "Dull 200 Results",
     order = c("Col", "19.8", "19.9"),
     labels = TRUE)}

}
